Add 'progress' to 'result'
I have a use case on a real live project to report both 
page progress and score for a SCORM-style course. I imagine 
this is a common use case. I'm surprised this hasn't come up 
before actually. 

I think the solution proposed is quite neat and simple. I don't
think an extension would be appropriate for something so common.