restructured section 4.1.10 Voided
As in my previous merges, I restructured the section for added clarity.
Please check for completeness and correctness. Note that I changed
'should' into 'must' in the original sentence that read: 'should mark
the target statement as voided using the "voided" field.' (line 1226)

I.m.h.o a core functionality as this is not optional. By all means,
change back to SHOULD if MUST is too stringent; I did not follow all
discussion around this.