Rewrite of 4.1.3 Verb
As a sort of ‘clean slate decision' I left out references to previous
versions of the spec. I feel this has only a historic meaning to the
limited number of people that are aware of / were involved in previous
versions of the spec. In anyone feels this should be included, leave a
comment.

Possible confusion: The description of the ID property says a URI should
be human-readable. Yet, "tag:example.com,2012:xQr73H" is given as an URI
example a little earlier and that is not very human-readable. Is this a
contradiction?