Rewrite of 4.1.11 Attachments
Well-structured original to begin with, but this section also contained
many sentences that were either descriptions, procedures or a
combination of both.(For example, the paragraph 'Transmission Format' in
the original). I took a stricter approach in separating that.

The Rationale-paragragh is also aimed at non-techies and as such
contained much too complex content. This was placed elsewhere in the
section.

Took out he unclear term "common requirements", this does not define
well what the requirements relate to (statements, attachments, LRS,
client?) and put those under the header "requirements for statement
streams that include attachments".

Made a literal copy-n-paste of the code examples, but they didn't look
good when previewing. Sorry if they turn out to be wrong on Github.
Someone more experienced with Github should please correct this.