Partial rewrite of 4.1.4 Object
subsection 4.1.4.1 to follow. Was not changed because too complex. Need
some co-writing done on my own fork before commit can be made.

Notable differences in this one: switched the order of StatementRef and
Sub-Statements. Made more sense while writing the 'Rationale' and I
expect StatementRef to be used more frequently.