Made Changes to the 3.0 Definitions Section
Fixed typos in TCAPI,MUST/SHOULD/MAY, and IFI tags and a couple redundancies in definitions.  

Renamed Learning Activity Provider to just Activity Provider as we don't have a common pattern for learning experiences, etc. throughout.  

Shortened some definitions to lessen "fluff".  The sections themselves provide more information.  

Re-defined activity to move away from "thing".  Confusing in the definition and never clarified later.  Used with Verb and Actor, which are introduced.

Updated URI to IRI.  Kept language.  I'm not as familiar with IRIs, so verify the meaning still works.

Completely re-defined LMS.

Removed "State" definition as it was the only API mentioned.  

Removed references to sections as it seems backwards for the definitions to refer to the document.