Accept 1.0 as 1.0.0
The spec has been talked about as 1.0 for a long time, and we've been shooting for (and succeeding) to allow 0.95 clients to continue to send statements with only an updated version header -- but we'd previously talked about updating it to 1.0.

If even a few pieces of content out there have done tried already (which they could have done before any 1.0(.0) systems exist by trying the 1.0 version header and falling back to 0.95 if rejected), it will be worth LRSs accepting 1.0 as 1.0.0. I don't see a downside, if one were to map 1.0 onto our new version scheme 1.0.0 is a logical equivalent.