Section 7.5 Edits
Please note the language change which referenced a DELETE method that
didn't exist.  An activity profile can only be deleted if both the
activity id and profile id exist.  This was ADL's interpretation, please
let us know if it was not the common one.  The old language referenced
this possibility but there was no corresponding method.