Language around Not rejecting Voiding Statements
The current version of the spec is silent on the behavior on what an LRS does if a voiding Statement comes in and the target is not available.  As we've discussed, timing issues could cause this delay and an LRS cannot reject solely on this.  This is a SHOULD NOT for this version and will be a MUST NOT for 2.0.