Miscellaneous Version Changes
We are now in 1.0.3, so I've updated this in some places, will do a sweeping change for the rest upon confirmation

Felt that LRS MUST accept may leave some cases that an LRS MUST NOT reject was clearer about (picky, I know, but seemed more accurate)

We can't enforce a description will be meaningful, so I changed that part to a SHOULD

Header language wasn't disambiguating name from value, so I added the value to the corresponding tag

Finally, I don't think we care too much if systems convert between the 1.0.Xs, what we care about are breaking changes i.e. minor and major versions.