Update Sections 4.1.7 and 4.1.8
I think most of this is pretty straightforward.  I just moved information which was in the table in 4.1 to these sections.  The one bullet that needs a look at is 

An LRS MUST ignore any value the Statement's stored property has when handling requests.

While the spec says the stored property is set by the LRS, we don't actually say what happens if it is used by an AP, etc.  I'm thinking we can't flat-out reject any Statement with "stored" because we expect LRS to LRS Statement transfers.  Ignoring makes sense, especially because "stored" is updated when transferred to a new system (per Appendix D).