IRI Fixes
Overall I think the 1.0.3Restructuring changes took care of most of the issue of #731, most of what was left was to expand Part 1's Community of Practice on Verbs and move any of the language if necessary. 

Please look over this change carefully as I've tried to retain as much of the meaning and original text as possible, but also to expand it to all types.  I didn't feel the need to list all types, in case that other expansions are made.  The point is to capture all "coinable" aka in a vocabulary data.