Timestamps and Stored
Per issues 719 and 733, moving Timestamp and stored to be in conjunction.  Also, note the type in the table changed from Date/Time to Timestamp to be consistent with the other Timestamp data types.  The requirements were also simplified.

The following two were the most concerning, the key question being can we assert these for ALL Timestamps, or just for those in timestamp and stored?  (this is probably a good question for any Timestamp as shown in Part 2 Section 4.5)

* The LRS SHOULD* return the Timestamp in UTC timezone.
* An LRS MUST NOT reject a Timestamp for being from the future, to prevent issues due to clock errors.