Fix ETag section for Transfer-Encoding
Resolves #789 and removes a small bit of erroneous language around ETag responses.

Specifically, this change clarifies that the calculated ETag only applies to responses without Transfer-Encoding (as that was the only way possible to respond with that rule in place), but opens the door for responses that use Transfer-Encoding and requires them to use a different ETag. This is a non-breaking change because all previously legal responses remain legal, and the additional responses possible are well-covered by internet standards (and would only break clients with basically broken implementations).

Surrounding that change, some starred SHOULDs around the calculated value are changed to non-starred, as they will no longer apply at some point in the future, rather than becoming requirements, and a star is added to the SHOULD about clients relying on the LRS-returned ETags.

Additionally, while reviewing the behavior of ETags in practice, I found the use of ETags in PUT and POST responses that do not include the entity body (as no LRS PUT/POST responses do) can cause problems with web infrastructure, particularly caching proxies. Accordingly, this PR removes the suggestion for LRSs to include the ETag there, since it is in conflict with web standards.